# docs/openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Literature Review Tool API",
    "description": "API for retrieving BibTeX data from academic databases",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/bibtex/scopus": {
      "get": {
        "summary": "Get Scopus Bibtex",
        "operationId": "get_scopus_bibtex_bibtex_scopus_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query for Scopus",
              "title": "Query"
            },
            "description": "Search query for Scopus"
          },
          {
            "name": "keywords",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated keywords to include in BibTeX",
              "default": "",
              "title": "Keywords"
            },
            "description": "Comma-separated keywords to include in BibTeX"
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of results",
              "default": 25,
              "title": "Count"
            },
            "description": "Number of results"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibTeXResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bibtex/springer": {
      "get": {
        "summary": "Get Springer Bibtex",
        "operationId": "get_springer_bibtex_bibtex_springer_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query for Springer",
              "title": "Query"
            },
            "description": "Search query for Springer"
          },
          {
            "name": "keywords",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated keywords to include in BibTeX",
              "default": "",
              "title": "Keywords"
            },
            "description": "Comma-separated keywords to include in BibTeX"
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of results",
              "default": 25,
              "title": "Count"
            },
            "description": "Number of results"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibTeXResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/bibtex/ieee": {
      "get": {
        "summary": "Get Ieee Bibtex",
        "operationId": "get_ieee_bibtex_bibtex_ieee_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query for IEEE",
              "title": "Query"
            },
            "description": "Search query for IEEE"
          },
          {
            "name": "keywords",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated keywords to include in BibTeX",
              "default": "",
              "title": "Keywords"
            },
            "description": "Comma-separated keywords to include in BibTeX"
          },
          {
            "name": "max_records",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of results",
              "default": 25,
              "title": "Max Records"
            },
            "description": "Number of results"
          },
          {
            "name": "start_record",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Start index for pagination",
              "default": 1,
              "title": "Start Record"
            },
            "description": "Start index for pagination"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibTeXResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BibTeXResponse": {
        "properties": {
          "total_results": {
            "type": "integer",
            "title": "Total Results"
          },
          "bibtex": {
            "type": "string",
            "title": "Bibtex"
          }
        },
        "type": "object",
        "required": [
          "total_results",
          "bibtex"
        ],
        "title": "BibTeXResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "type": "object",
        "required": ["error"]
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "APIKeyHeader": []
    }
  ]
}